+++ core-ng/src/main/java/core/framework/internal/db/DatabaseImpl.java
@@ -311,7 +311,7 @@ public final class DatabaseImpl implements Database {
         ActionLog actionLog = LogManager.currentActionLog();
         if (actionLog != null) {
             actionLog.stats.compute("db_queries", (k, oldValue) -> (oldValue == null) ? queries : oldValue + queries);
-            actionLog.track("db", elapsed, readRows, writeRows);
+            actionLog.track("db", elapsed, readRows, writeRows, 0, 0);
         }
     }