+++ core-ng/src/main/java/core/framework/internal/web/sse/ChannelSupport.java
@@ -1,5 +1,8 @@
 package core.framework.internal.web.sse;
 
+import core.framework.internal.log.PerformanceWarning;
+import core.framework.internal.log.WarningContext;
+import core.framework.log.IOWarning;
 import core.framework.web.Request;
 import core.framework.web.rate.LimitRate;
 import core.framework.web.sse.Channel;
@@ -14,24 +17,27 @@ class ChannelSupport<T> {
     final ServerSentEventWriter<T> builder;
     @Nullable
     final LimitRate limitRate;
+    final PerformanceWarning @Nullable [] warnings;
 
     ChannelSupport(ChannelListener<T> listener, Class<T> eventClass, ServerSentEventContextImpl<T> context) {
         this.listener = listener;
         this.context = context;
         builder = new ServerSentEventWriter<>(eventClass);
-        limitRate = limitRate(listener);
-    }
 
-    @Nullable
-    private LimitRate limitRate(ChannelListener<T> listener) {
         try {
             Method targetMethod = listener.getClass().getMethod("onConnect", Request.class, Channel.class, String.class);
-            LimitRate limitRate = targetMethod.getDeclaredAnnotation(LimitRate.class);
-            if (limitRate == null)
-                limitRate = listener.getClass().getDeclaredAnnotation(LimitRate.class);
-            return limitRate;
+            limitRate = limitRate(targetMethod);
+            warnings = WarningContext.warnings(targetMethod.getDeclaredAnnotationsByType(IOWarning.class));
         } catch (NoSuchMethodException e) {
             throw new Error("failed to get listener.onConnect method", e);
         }
+    }
+
+    @Nullable
+    private LimitRate limitRate(Method targetMethod) {
+        LimitRate limitRate = targetMethod.getDeclaredAnnotation(LimitRate.class);
+        if (limitRate == null)
+            limitRate = listener.getClass().getDeclaredAnnotation(LimitRate.class);
+        return limitRate;
     }
 }