+++ core-ng/src/main/java/core/framework/log/ActionLogContext.java
@@ -50,11 +50,15 @@ public final class ActionLogContext {
         return track(operation, elapsed, 0, 0);
     }
 
-    // return the total count of operations within current action
     public static int track(String operation, long elapsed, int readEntries, int writeEntries) {
+        return track(operation, elapsed, readEntries, writeEntries, 0, 0);
+    }
+
+    // return the total count of operations within current action
+    public static int track(String operation, long elapsed, int readEntries, int writeEntries, long readBytes, long writeBytes) {
         ActionLog actionLog = LogManager.currentActionLog();
         if (actionLog == null) return 1;    // be called without action context
-        return actionLog.track(operation, elapsed, readEntries, writeEntries);
+        return actionLog.track(operation, elapsed, readEntries, writeEntries, readBytes, writeBytes);
     }
 
     public static void triggerTrace(boolean cascade) {