+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### 9.3.1 (10/8/2025 - )
 
+* tracking: add readBytes/writeBytes to perf_test, to separate read row/doc/keys and read bytes
+* cache: track redis store read/write bytes and elapsed time (included json serialization/deserialization time)
+* http: unified http slow threshold config, use @IOWarning to config for each action
+* sse: support @IOWarning on handler
+* http: update okhttp to 5.2.1
 
 ### 9.3.0 (8/27/2025 - 10/8/2025) !!! only support java 25